+++ test/specs/repo/extensions/syntax-highlight.ts
@@ -1,7 +1,7 @@
 import { readFile } from 'fs/promises';
 import { skipIfExtensionNotPresent } from '../../../helpers/default-functions.js';
 import { utf8 } from '../../../helpers/readFileEncoding.js';
-import awaitDisplayed from '../../../helpers/awaitDisplayed.js';
+import awaitDisplayed from '../../../helpers/await-displayed.js';
 
 describe( 'SyntaxHighlight', function () {
 	beforeEach( async function () {
